Add methods for mapping specific types
This adds methods for mapping double, float, int and string.
When mapping a type, if it cannot be resolved directly, it will now try
to resolve it as the closest type available.

If you map a Float but get a Double, Tailor will try to make a Float
out of the Double before returning.